fix: adapt to Long id

+++ src/main/kotlin/es/uib/tfg/sports_league_backend/common/security/JwtService.kt
@@ -18,7 +18,7 @@ class JwtService(
         return Keys.hmacShaKeyFor(secretKey.toByteArray())
     }
 
-    fun generateToken(userId: Int, type: TokenType): String {
+    fun generateToken(userId: Long, type: TokenType): String {
         val tokenExpiration = when (type) {
             TokenType.ACCESS -> jwtExpiration
             TokenType.REFRESH -> jwtRefreshExpiration
